Optimization of RSC and RFC client-side to maximize load times. Added Google Maps component.
