+++ src/components/MainViewComponents/Home.tsx
@@ -1,17 +1,10 @@
-'use client'
-
 import React from 'react'
 import Image from 'next/image'
 import { Button } from '../ReUsableComponents/Button'
 
 export const Home = () => {
 
-  const myOnclick = function () {
-    window.open(
-      'https://wa.me/+50255643166?text=Hola%21%0AQuisiera%20cotizar%20repuestos.',
-      '_blank' // <- This is what makes it open in a new window.
-    );
-  }
+  const wpLink = "https://wa.me/+50255643166?text=Hola%21%0AQuisiera%20cotizar%20repuestos."
 
   return (
     <div className='relative flex'>
@@ -26,7 +19,8 @@ export const Home = () => {
       <Button
         text='Cotizar mi repuesto'
         className='bottom-[25%] left-[8%] absolute bg-hc-blue text-slate-100 hover:scale-125 duration-200 font-bold tracking-wider md:text-lg text-sm px-2'
-        onClick={myOnclick}
+        type='whatsapp'
+        wpMessage={wpLink}
       />
     </div>
   )
