+++ src/components/MainViewComponents/Locations.tsx
@@ -1,26 +1,11 @@
-'use client'
-
-import React, { useEffect } from 'react'
+import React from 'react'
 import { Button } from '../ReUsableComponents/Button';
 import LocationElement from '../ReUsableComponents/LocationElement';
+import GoogleMap from '../ReUsableComponents/GoogleMap';
 
 export const Locations = () => {
 
-  let map: google.maps.Map;
-  const center: google.maps.LatLngLiteral = { lat: 14.6472287, lng: -90.5254478 };
-
-  // Markers
-  const markerPosition = [
-    { 'postition': { lat: 14.6472287, lng: -90.5254478 }, 'label': 'Periférico' },
-    { 'postition': { lat: 14.6195206, lng: -90.5468998 }, 'label': 'Calz. San Juan' },
-    { 'postition': { lat: 14.6672277, lng: -90.8494999 }, 'label': 'Chimaltenango' },
-    { 'postition': { lat: 14.3117942, lng: -90.3750463 }, 'label': 'Barberena' },
-    { 'postition': { lat: 14.2990816, lng: -90.7841798 }, 'label': 'Escuintla' },
-    { 'postition': { lat: 14.8405614, lng: -91.5015162 }, 'label': 'Xela' },
-    { 'postition': { lat: 14.9925009, lng: -89.7201585 }, 'label': 'Teculután' },
-    { 'postition': { lat: 15.3010977, lng: -91.4963686 }, 'label': 'Huehuetenango' },
-    { 'postition': { lat: 14.6469972, lng: -90.5257205 }, 'label': 'Centro de Servicio' }
-  ];
+  const wpMessage = "https://cophershop.com/"
 
   // LocationsElements
   const locations = [
@@ -34,36 +19,6 @@ export const Locations = () => {
     {"title": "Huehuetenango", "adress": "Sector6, Cambote zona 11, Huehuetenango"},
     {"title": "Centro de Servicio", "adress": "Anillo Periferico 4-10 zona 3, Guatemala"}
   ]
-
-  function initMap(): void {
-    map = new google.maps.Map(document.getElementById("map") as HTMLElement, {
-      center,
-      zoom: 13
-    });
-
-    // The marker, positioned at HC locations
-    markerPosition.map((mark, index) => {
-      const label = mark.label
-      const position = mark.postition
-
-      const marker = new google.maps.Marker({
-        map: map,
-        position: position,
-        title: label
-      })
-    })
-  }
-
-  const myOnclick = function () {
-    window.open(
-      'https://cophershop.com/',
-      '_blank' // <- This is what makes it open in a new window.
-    );
-  }
-
-  useEffect(() => {
-    initMap()
-  })
 
   return (
     <div className='flex flex-col w-full' id="ubicaciones">
@@ -84,12 +39,13 @@ export const Locations = () => {
             <Button
               text='Tiendas'
               className='bg-hc-red m-auto hover:scale-125 duration-200 font-bold text-xl tracking-wider outline outline-1 outline-offset-2 outline-white px-2'
-              onClick={myOnclick}
+              type='whatsapp'
+              wpMessage={wpMessage}
             />
           </div>
         </div>
         <div className='w-full h-full max-w-full'>
-          <div id="map" className='w-full h-full'></div>
+          <GoogleMap />
         </div>
       </div>
       <div className='bg-hc-red max-h-[40vh] md:h-[20vh] grid grid-cols-2 md:grid-cols-5 gap-3 py-4 px-2 overflow-y-auto'>
