+++ src/components/MainViewComponents/Services.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import Image from 'next/image'
 import React from 'react'
